requirements.txt
python-telegram-bot
deep-translator
yake